pyproject.toml
[project]
name = "bids-guidelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docopt>=0.6.2",
    "pybids>=0.19.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "datalad>=1.2.1",
    "datalad-installer>=1.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
]
